chore(CHANGELOG): add changelogs for v2.0.0 (#35)
Along with Changelog updates, this commit fixes the travis
 - bumps the k8s version in minikube
 - updates the latest cstor operator
   yaml with new CRDs
 - improves error reporting by adding logs to tests

+++ ci/upgrade/sanity.sh
@@ -38,6 +38,11 @@ export TEST_VERSION=${TEST_VERSION#v}
 # setup openebs & cstor v1 for migration 
 ./ci/upgrade/setup.sh || exit 1
 # run migration tests
-./ci/upgrade/test.sh || exit 1
+./ci/upgrade/test.sh 
+if [[ $? != 0 ]]; then
+  kubectl logs --tail=50 -l job-name=upgrade-pool -n openebs
+  kubectl logs --tail=50 -l job-name=upgrade-volume -n openebs
+  exit 1
+fi
 
 rm ./ci/upgrade/volume.yaml ./ci/upgrade/application.yaml
